+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/fragments/MyAccountTabFragment.java
@@ -71,6 +71,8 @@ public class MyAccountTabFragment extends Fragment {
         }
     };
 
+    public  MyAccountTabFragment(){}
+
     @SuppressLint("ValidFragment")
     public MyAccountTabFragment(int iOpenTabPosition) {
         iLastTabPosition = iOpenTabPosition;