+++ Guildford/MHSystems/CHANGELOG.md
@@ -27,6 +27,7 @@
     
     ####16th March, 2017
      - Fixed: Resolved crash of wrong menu item displayed on YOUR ACCOUNT option.
+     - Fixed: Resolved crashed. java.lang.Class<com.mh.systems.guildford.fragments.MyAccountTabFragment> has no zero argument constructor
 
 ##Version 1.2.0 (2016-12-26)